+++ src/Braintree/DisputeTransaction.cs
@@ -7,6 +7,7 @@ namespace Braintree
         public virtual decimal? Amount { get; protected set; }
         public virtual DateTime? CreatedAt { get; protected set; }
         public virtual string Id { get; protected set; }
+        public virtual int? InstallmentCount { get; protected set; }
         public virtual string OrderId { get; protected set; }
         public virtual string PaymentInstrumentSubtype { get; protected set; }
         public virtual string PurchaseOrderNumber { get; protected set; }
@@ -16,6 +17,7 @@ namespace Braintree
             Amount = node.GetDecimal("amount");
             CreatedAt = node.GetDateTime("created-at");
             Id = node.GetString("id");
+            InstallmentCount = node.GetInteger("installment-count");
             OrderId = node.GetString("order-id");
             PaymentInstrumentSubtype = node.GetString("payment-instrument-subtype");
             PurchaseOrderNumber = node.GetString("purchase-order-number");